Protocol: move Protocol trait to protocols.rs
move Protocol trait from lib.rs to protocols.rs

fix Protocol::guid() return type from borrow value to const raw pointer
    - if using borrow value, handle_protocol() doesn't not work.
